chore: move fs-utils file to hmr test project
